requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
dnspython==2.6.1
httpx==0.27.2
pyyaml==6.0.2
